Fix - change map value from size_t to uint32_t Add - add struct resource name in order to link resource_t to a string name Add - look function -> player look around him depending on his level Add - look and print_content files and headers related to look functions

+++ include/server/map.h
@@ -38,8 +38,8 @@ typedef struct cell_s {
 } cell_t;
 
 typedef struct map_s {
-	size_t x;
-	size_t y;
+	uint32_t x;
+	uint32_t y;
 	cell_t **map;
 } map_t;
